+++ Common/src/main/java/fuzs/universalenchants/config/ServerConfig.java
@@ -32,7 +32,7 @@ public class ServerConfig implements ConfigCore {
     public boolean noFarmlandTrample = true;
     @Config(category = "enchantment_improvements", description = "Looting also affects dropped experience points.")
     public boolean lootingBoostsXp = true;
-    @Config(category = "enchantment_improvements", description = {"Remove the max level cap from the /enchant command, also allow overriding existing enchantment levels.", "Additionally make enchanting books work via the command."})
+    @Config(category = "enchantment_improvements", description = {"Remove the max level cap from the /enchant command, also allow overriding and removing (by setting the level to 0) existing enchantment levels.", "Additionally make enchanting books work via the command."})
     public boolean fixEnchantCommand = true;
 
     public static class ItemCompatibilityConfig implements ConfigCore {